IGNITE-20042 / rename a class
