+++ modules/core/src/main/java/org/apache/ignite/internal/Kludges.java
@@ -18,9 +18,9 @@
 package org.apache.ignite.internal;
 
 /**
- * Contains hacks needed for the whole codebase. Should be removed as quickly as possible.
+ * Contains kludges needed for the whole codebase. Should be removed as quickly as possible.
  */
-public class Hacks {
+public class Kludges {
     // TODO: Remove after IGNITE-20499 is fixed.
     /** Name of the property overriding idle safe time propagation period (in milliseconds). */
     public static final String IDLE_SAFE_TIME_PROPAGATION_PERIOD_MILLISECONDS_PROPERTY = "IDLE_SAFE_TIME_PROPAGATION_PERIOD_MILLISECONDS";
